requirements.txt
# requirements.txt

# Core Django
django>=4.2 # Or your preferred Django version

# WSGI server for production (Gunicorn is used in Dockerfile/compose examples)
gunicorn

# PostgreSQL adapter
psycopg2-binary

# Environment variable management
django-environ # For reading .env files and managing settings

# Django REST Framework and related packages
djangorestframework>=3.14 # For building REST APIs
django-cors-headers # For handling Cross-Origin Resource Sharing (CORS)
drf-spectacular # For generating OpenAPI 3 schema and Swagger/ReDoc documentation

# Image handling (if you use ImageField)
Pillow

# Static files serving in production (if not using a separate web server like Nginx for static files)
whitenoise

# Optional: HTTP client (if your app needs to make external API calls)
# requests

# Add any other Python packages your project needs, for example:
# celery # For asynchronous tasks
# redis # Often used with Celery as a message broker
# python-dotenv # Generally not needed if using django-environ and Docker env_file, but sometimes used for local non-Docker scripts


redis==5.2.1
python-dotenv
google-generativeai==0.8.3
celery==5.5.2
djangorestframework-simplejwt
# For Google Cloud Storage integration
google-auth==2.37.0

# For Google Cloud AI Platform integration
google-cloud-aiplatform==1.81.0
#An integration package connecting Google VertexAI and LangChain
langchain==0.3.16
langchain-google-vertexai==2.0.13
google-cloud-bigquery==3.29.0
google-cloud-datastore==2.20.2
google-cloud-speech==2.29.0
google-cloud-texttospeech==2.23.0
pydub==0.25.1
sounddevice==0.5.1
soundfile==0.12.1